chore(Jest): setup migration (algolia/algoliasearch-helper-js#659)
* chore(Jest): functions + hierarchical-facets [PART-1] (algolia/algoliasearch-helper-js#660)

* chore(Jest): SearchParameters [PART-2] (algolia/algoliasearch-helper-js#661)

* chore(Jest): SearchResults [PART-3] (algolia/algoliasearch-helper-js#662)

* chore(Jest): algoliasearch.helper [PART-4] (algolia/algoliasearch-helper-js#663)

* chore(Jest): top level spec [PART-5] (algolia/algoliasearch-helper-js#664)

* chore(Jest): integration [PART-6] (algolia/algoliasearch-helper-js#665)

* chore(Jest): runner [PART-7] (algolia/algoliasearch-helper-js#666)

* chore(Jest): cleanup [PART-8] (algolia/algoliasearch-helper-js#667)

+++ packages/algoliasearch-helper/test/run.js
@@ -1,26 +1,42 @@
 'use strict';
 
-var bulk = require('bulk-require');
-var test = require('tape');
-var algolia = require('algoliasearch');
+var path = require('path');
+var jest = require('jest');
+var algoliasearch = require('algoliasearch');
+var staticJestConfig = require('../jest.config');
 
-bulk(__dirname, ['spec/**/*.js']);
+var enableIntegrationTest =
+  process.env.ONLY_UNIT !== 'true' &&
+  process.env.INTEGRATION_TEST_API_KEY &&
+  process.env.INTEGRATION_TEST_APPID;
 
-if (process.env.ONLY_UNIT !== 'true' && process.env.INTEGRATION_TEST_API_KEY && process.env.INTEGRATION_TEST_APPID) {
-  // usage: INTEGRATION_TEST_APPID=$APPID INTEGRATION_TEST_API_KEY=$APIKEY npm run dev
-  bulk(__dirname, ['integration-spec/**/*.js']);
+var projectsRootPaths = [path.resolve(__dirname, '..')];
+var dynamicJestConfig = Object.assign({}, staticJestConfig, {
+  maxWorkers: 4,
+  setupFilesAfterEnv: staticJestConfig.setupFilesAfterEnv || []
+});
 
-  test.onFinish(cleanupIntegration);
+if (enableIntegrationTest) {
+  dynamicJestConfig.testMatch.push('<rootDir>/test/integration-spec/**/*.[jt]s?(x)');
+  dynamicJestConfig.setupFilesAfterEnv.push(path.resolve(__dirname, '..', 'jest.setup.js'));
 }
 
-function cleanupIntegration() {
-  console.log('Deleting all indices');
-  var client = algolia(process.env.INTEGRATION_TEST_APPID, process.env.INTEGRATION_TEST_API_KEY);
-  var maybeIndices = client.listIndexes();
-  maybeIndices.then(content => {
-    content.items
-      .map(i => i.name)
-      .filter(n => n.indexOf('_travis-algoliasearch-helper') !== -1)
-      .forEach(n => client.deleteIndex(n));
-  });
-}
+jest.runCLI(dynamicJestConfig, projectsRootPaths).then(function(response) {
+  if (!response.results.success) {
+    process.exitCode = response.globalConfig.testFailureExitCode;
+  }
+
+  if (enableIntegrationTest) {
+    var client = algoliasearch(
+      process.env.INTEGRATION_TEST_APPID,
+      process.env.INTEGRATION_TEST_API_KEY
+    );
+
+    client.listIndexes().then(content => {
+      content.items
+        .map(i => i.name)
+        .filter(n => n.indexOf('_travis-algoliasearch-helper') !== -1)
+        .forEach(n => client.deleteIndex(n));
+    });
+  }
+});
